+++ libzmq/zmq/bp_engine.hpp
@@ -23,6 +23,7 @@
 
 #include <string>
 
+#include <zmq/stdint.hpp>
 #include <zmq/export.hpp>
 #include <zmq/i_engine.hpp>
 #include <zmq/i_pollable.hpp>
@@ -63,7 +64,7 @@ namespace zmq
 
         //  i_pollable interface implementation.
         engine_type_t type ();
-        void get_watermarks (int *hwm_, int *lwm_);
+        void get_watermarks (uint64_t *hwm_, uint64_t *lwm_);
         void process_command (const engine_command_t &command_);
         void register_event (i_poller *poller_);
         void in_event ();
