+++ libzmq/zmq/in_engine.hpp
@@ -20,6 +20,7 @@
 #ifndef __ZMQ_IN_ENGINE_HPP_INCLUDED__
 #define __ZMQ_IN_ENGINE_HPP_INCLUDED__
 
+#include <zmq/stdint.hpp>
 #include <zmq/i_engine.hpp>
 #include <zmq/mux.hpp>
 
@@ -30,24 +31,24 @@ namespace zmq
     {
     public:
 
-        ZMQ_EXPORT static in_engine_t *create (int hwm_, int lwm_);
+        ZMQ_EXPORT static in_engine_t *create (uint64_t hwm_, uint64_t lwm_);
 
         ZMQ_EXPORT bool read (message_t *msg_);
 
         //  i_engine implementation.
         engine_type_t type ();
-        void get_watermarks (int *hwm_, int *lwm_);
+        void get_watermarks (uint64_t *hwm_, uint64_t *lwm_);
         void process_command (const engine_command_t &command_);
 
     private:
 
-        in_engine_t (int hwm_, int lwm_);
+        in_engine_t (uint64_t hwm_, uint64_t lwm_);
         ~in_engine_t ();
 
         mux_t mux;
 
-        int hwm;
-        int lwm;
+        uint64_t hwm;
+        uint64_t lwm;
     };
 
 }
